requirements.txt
# Transcription Pipeline - Python Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# Speaker Diarization (Local)
pyannote.audio>=3.1.1
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0

# Transcription Engines - Groq
requests>=2.31.0
openai>=1.0.0

# OpenAI Whisper (fallback)
openai-whisper>=20231117
faster-whisper>=0.10.0

# Document Output
python-docx>=1.1.0
markdown>=3.5.0

# File Watching
watchdog>=3.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0
sqlmodel>=0.0.14

# Utilities
tqdm>=4.66.0
python-magic>=0.4.27
httpx>=0.25.0
aiofiles>=23.2.0
rich>=13.7.0

# Queue & Caching
redis>=5.0.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0
slowapi>=0.1.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
